# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/search/query": {
      "post": {
        "operationId": "WebSearchController_search",
        "summary": "Execute web search (DuckDuckGo/fallback) and save query to database",
        "description": "\n### Service Architecture & Web Scraping Pipeline\n- **Zero-Dependency Live Scraping**: Dispatches outbound HTTP requests to DuckDuckGo's public HTML interface using a realistic User-Agent rotation strategy to bypass automated bot challenges.\n- **Cheerio DOM Parsing**: Parses the raw HTML response tree using Cheerio, extracting organic search results including sanitized clean URLs, snippet summaries, and page titles.\n- **Resilient Fallback Engine**: If DuckDuckGo triggers anti-scraping CAPTCHAs or timeouts, the service gracefully switches to a curated fallback result set or secondary public search mirror, ensuring uninterrupted user experience.\n- **Auditing & History Logging**: Asynchronously writes the query, user relationship, result count, and serialized top result summaries to the `WebSearch` PostgreSQL table for auditability and fast re-retrieval.\n\n### Future Scalability Roadmap\n- **Distributed Redis Search Cache**: Cache normalized search query results (`search:query:<hash>`) with a 60-minute TTL. Popular search queries (e.g. \"latest tech news\") resolve in <2ms directly from RAM without hitting external search engine endpoints.\n- **Asynchronous Crawling Worker Pool**: For advanced agentic search (fetching and summarizing full target webpage contents for RAG), delegate URL deep-fetching to a background worker queue (BullMQ + Redis) to maintain low HTTP request latency on the primary API gateway.\n    ",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchQueryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Search results parsed and returned successfully"
          },
          "400": {
            "description": "Bad Request - Search query empty or invalid"
          },
          "401": {
            "description": "Unauthorized - Missing or invalid Bearer JWT"
          }
        },
        "tags": [
          "Web Search API"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/search/history": {
      "get": {
        "operationId": "WebSearchController_getHistory",
        "summary": "Get paginated web search history of the current user",
        "description": "\n### Query Optimization & User Isolation\n- **Tenant Isolation**: Strictly scopes queries to the authenticated `userId` extracted from the validated JWT token.\n- **B-Tree Indexed Pagination**: Leverages database indexes on `(userId, createdAt DESC)` to deliver sub-millisecond paginated responses even with hundreds of thousands of historical user search logs.\n    ",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (default: 1)",
            "schema": {
              "example": 1,
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Results per page (default: 20, max: 100)",
            "schema": {
              "example": 20,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated user search history"
          }
        },
        "tags": [
          "Web Search API"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/search/recent": {
      "get": {
        "operationId": "WebSearchController_getRecent",
        "summary": "Get recent distinct search queries for quick suggestion chips",
        "description": "\n### Service Flow\n- **Deduplicated Query Retrieval**: Aggregates the user's most recent distinct queries using a `DISTINCT ON (query)` SQL query ordered by `createdAt DESC`.\n- **Extension UI Integration**: Powers instant \"Recent Searches\" chips in the EchoGPT Chrome Extension popup for single-click search re-execution.\n    ",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of recent queries to return (default: 5)",
            "schema": {
              "example": 5,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of recent distinct query strings"
          }
        },
        "tags": [
          "Web Search API"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/search/suggestions": {
      "get": {
        "operationId": "WebSearchController_getSuggestions",
        "summary": "Get auto-complete search query suggestions",
        "description": "\n### Autocomplete Architecture\n- **Low-Latency Search Assist**: Queries upstream auto-complete endpoints and local historical query indices to return instant search phrase completions as the user types in the EchoGPT extension search bar.\n- **Debounce Tolerance**: Optimized for high-concurrency debounced client keystroke queries.\n\n### Future Scalability Roadmap\n- **Trie / Inverted Index in Redis**: In high-scale deployments, maintain a distributed prefix-tree (Trie) in Redis for instant O(k) prefix matching across millions of platform queries without database roundtrips.\n    ",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "description": "Partial query prefix string",
            "schema": {
              "example": "NestJS",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of suggested query completions"
          }
        },
        "tags": [
          "Web Search API"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "EchoGPT REST API — Architecture & Engineering Specification",
...
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Provide your Bearer Access Token (issued from /auth/login or /auth/register)",
        "in": "header"
      }
    },
    "schemas": {
...
      "SearchQueryDto": {
        "type": "object",
        "properties": {
          "query": {
            "type": "string",
            "description": "Search terms or question to look up"
          },
          "maxResults": {
            "type": "number",
            "description": "Maximum search results to return (1-20)",
            "example": 5,
            "default": 5,
            "minimum": 1,
            "maximum": 20
          }
        },
        "required": [
          "query"
        ]
      },
...
    }
  }
}